refactor: split four oversized pages into colocated components (#130)
Page sizes:
  apartments/[id]/page.tsx  643 → 440 lines (-32%)
  apartments/new/page.tsx   626 → 376 lines (-40%)
  apartments/page.tsx       501 → 365 lines (-27%)
  compare/page.tsx          535 → 209 lines (-61%)

Three of four under the 400-line target. apartments/[id] sits at 440
because the remaining mass is state-coupled handlers (delete,
reprocess, edit, save-rating) that don't extract cleanly without a
larger custom-hook refactor — saving that for a follow-up.

New colocated components:
  src/app/apartments/_components/
    apartment-summary.ts, apartment-badges.tsx,
    apartment-card.tsx, apartment-row.tsx
  src/app/apartments/[id]/_components/
    types.ts, distance-section.tsx, apartment-pager-nav.tsx,
    apartment-actions.tsx, apartment-metric-badges.tsx
  src/app/apartments/new/_components/
    types.ts, status-badge.tsx, upload-step.tsx,
    review-step.tsx, single-entry-step.tsx
  src/app/compare/_components/
    compare-types.ts, compare-column-header.tsx, compare-table.tsx

All existing tests pass (338/338). Removed an unused helper from
the new full-flow.test.tsx that lint flagged after the refactor.
